+++ packages/aspida-swagger/samples/jsonV3/api/v1/mixes/_mixId/url.ts
@@ -1,0 +1,10 @@
+/* eslint-disable */
+export interface Methods {
+  post: {
+    reqData: {
+      url: string
+      published: boolean
+      enableImage: boolean
+    }
+  }
+}