+++ packages/aspida-swagger/samples/jsonV3/api/v1/mixes/index.ts
@@ -1,0 +1,21 @@
+/* eslint-disable */
+export interface Methods {
+  get: {
+    resData: {
+      id: number
+      title: number
+      createdByCurrentClientAPI: boolean
+    }[]
+  }
+
+  post: {
+    resData: {
+      id: number
+    }
+
+    reqData: {
+      name: string
+      title: string
+    }
+  }
+}