+++ packages/aspida-swagger/samples/jsonV3/api/v3/channels/_channelId/chats/_chatId/items/_itemId.ts
@@ -1,0 +1,16 @@
+/* eslint-disable */
+import * as Types from '../../../../../../../@types'
+
+export interface Methods {
+  post: {
+    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader
+
+    resData: {
+      ok: boolean
+    }
+
+    reqData: {
+      formattedText: string
+    }
+  }
+}