+++ packages/aspida-swagger/samples/jsonV3/api/v3/chats/keys.ts
@@ -1,0 +1,19 @@
+/* eslint-disable */
+import * as Types from '../../../@types'
+
+export interface Methods {
+  get: {
+    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken & Types.UserPublicKey & Types.UserSignedChallenge
+
+    resData: {
+      token: string
+      publishKey: string
+      subscribeKey: string
+      cipherKey: string
+      senderDevice: {
+        ios: string[]
+        android: string[]
+      }
+    }
+  }
+}