+++ packages/aspida-swagger/samples/jsonV3/api/v3/extension/story/_storyId/video.ts
@@ -1,0 +1,22 @@
+/* eslint-disable */
+import * as Types from '../../../../../@types'
+
+export interface Methods {
+  post: {
+    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken
+
+    resData: {
+      id: string
+    }
+
+    reqType: FormData
+
+    reqData: {
+      file: ArrayBuffer
+      caption: string
+      headline: string
+      text: string
+      published: boolean
+    }
+  }
+}