+++ packages/aspida-swagger/samples/jsonV3/api/v3/organisation/index.ts
@@ -1,0 +1,10 @@
+/* eslint-disable */
+import * as Types from '../../../@types'
+
+export interface Methods {
+  get: {
+    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationTokenRequired
+
+    resData: Types.OrganisationModel
+  }
+}