+++ packages/aspida-swagger/samples/jsonV3/api/v3/organisation/users.ts
@@ -1,0 +1,26 @@
+/* eslint-disable */
+import * as Types from '../../../@types'
+
+export interface Methods {
+  get: {
+    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationTokenRequired
+
+    query: {
+      search: string
+      offset: number
+    }
+
+    resData: {
+      count: number
+      offset: number
+      data: {
+        email: string
+        id: number
+        initials: string
+        role: string
+        roleScope: string
+        screenName: string
+      }[]
+    }
+  }
+}