+++ packages/aspida-swagger/samples/jsonV3/api/v3/stories/_storyId/items/index.ts
@@ -1,0 +1,18 @@
+/* eslint-disable */
+import * as Types from '../../../../../@types'
+
+export interface Methods {
+  get: {
+    reqHeaders: Types.AppIdHeader & Types.AppPlatformHeader & Types.AppVersionHeader & Types.AppOrganisationToken & Types.UserInstanceIdHeader
+
+    query: {
+      offset: number
+    }
+
+    resData: {
+      limit: number
+      offset: number
+      data: Types.ModelCard[]
+    }
+  }
+}