+++ packages/aspida-swagger/samples/yamlV2/pet/_petId@number/index.ts
@@ -1,0 +1,11 @@
+/* eslint-disable */
+export interface Methods {
+  get: {
+  }
+
+  post: {
+  }
+
+  delete: {
+  }
+}