+++ packages/aspida-swagger/samples/yamlV2/store/order/_orderId@number.ts
@@ -1,0 +1,8 @@
+/* eslint-disable */
+export interface Methods {
+  get: {
+  }
+
+  delete: {
+  }
+}