+++ packages/aspida/src/cli.ts
@@ -1,22 +1,21 @@
 import fs from 'fs'
 import minimist from 'minimist'
-import getConfig, { Config } from './getConfig'
+import getConfig, { BaseConfig } from './getConfig'
 import write from './writeRouteFile'
 import watch from './watchInputDir'
 import { Build, Watch, CommandToBuild } from './cli/build'
 import { Command, nullCommand } from './cli/command'
 import { version as versionCommand } from './cli/version'
 
-const options: minimist.Opts = {
+export const options: minimist.Opts = {
   string: ['version', 'config', 'build', 'watch'],
   alias: { v: 'version', c: 'config', b: 'build', w: 'watch' }
 }
 
-const getBuildCommandFactory = (config: Config) =>
+const getBuildCommandFactory = (config: BaseConfig[]) =>
   CommandToBuild.getFactory(config, {
     write,
     watch,
-    read: ({ input }) => (Array.isArray(input) ? input : [input]),
     remove: (filePath: string, callback: () => void) => fs.unlink(filePath, callback)
   })