+++ packages/aspida/src/cli/build.ts
@@ -1,13 +1,13 @@
-import { Config } from '../getConfig'
+import { BaseConfig } from '../getConfig'
 import build from '../buildTemplate'
 import { Template } from '../build/template'
 import { Command } from './command'
 
 export class CommandToBuild implements Command {
-  static getFactory(config: Config, io: BuildIO) {
+  static getFactory(configs: BaseConfig[], io: BuildIO) {
     return {
       create(command: BuildCommand): Command {
-        return new CommandToBuild(command, config, io)
+        return new CommandToBuild(command, configs, io)
       }
     }
   }
@@ -15,31 +15,30 @@ export class CommandToBuild implements Command {
   // eslint-disable-next-line no-useless-constructor
   private constructor(
     private readonly command: BuildCommand,
-    private readonly config: Config,
+    private readonly configs: BaseConfig[],
     private readonly io: BuildIO
   ) {}
 
   exec() {
-    this.io.read(this.config).forEach(input => {
-      this.command.run(input, this.io)
+    this.configs.forEach(config => {
+      this.command.run(config, this.io)
     })
   }
 }
 
 interface BuildCommand {
-  run(input: string, io: BuildIO): void
+  run(config: BaseConfig, io: BuildIO): void
 }
 
 export interface BuildIO {
-  read(config: Config): string[]
   write(template: Template): void
   remove(filePath: string, callback: () => void): void
   watch(input: string, callback: () => void): void
 }
 
 export class Build implements BuildCommand {
-  run(input: string, io: BuildIO): void {
-    const template = build(input)
+  run(config: BaseConfig, io: BuildIO): void {
+    const template = build(config)
 
     io.remove(template.filePath, () => io.write(template))
   }
@@ -49,8 +48,8 @@ export class Watch implements BuildCommand {
   // eslint-disable-next-line no-useless-constructor
   constructor(private readonly build = new Build()) {}
 
-  run(input: string, io: BuildIO): void {
-    this.build.run(input, io)
-    io.watch(input, () => this.build.run(input, io))
+  run(config: BaseConfig, io: BuildIO): void {
+    this.build.run(config, io)
+    io.watch(config.input, () => this.build.run(config, io))
   }
 }